Move Template to provider
